Empty List In NotThe Same As None
We have to be careful when we check for return values. An empty list is not the same as None. I just discovered this in get_primvars. There may be other cases within the code I have not detected. Be careful using the is not None technique. It only works if your calling def actually returns None.