stats.xml changes
Changed output of stats.xml to reference current frame (ex. "stats_100.xml"), along with "Open Last Stats" to load the stats file for the current frame you're on in Blender to allow for easy comparison of various frames, and changed the properties text of "Output stats" to better match how it currently is (since it's an .xml file, not a .txt file).