Improve logic for finding an rmantree
Even if the $RMANTREE is set to a valid rmantree, the addon would fail to
load if there's no /opt/pixar or if the rmantree doesn't contain
"RenderManProServer" somewhere in it's path.

This change guards against failure if there's no /opt/pixar (or analogous
directory on other platforms), determines the version by parsing
`$rmantree/bin/prman -version` instead of the path itself, and streamlines
the code to guess the newest installed rmantree.