Replace PRMan with RenderMan in user-visible strings
The current official style these days is to call the renderer simply
"RenderMan".  It's just the executable itself that is called prman.
Referring to the renderer as RenderMan where possible helps make the UI
more consistent and uniform, and a bit friendlier to less technical users
who might not know how prman relates to RenderMan.