Fix exported cache sizes
The cache sizes were being multiplied by 100 on export.  On an older,
down-spec, machine with only 12GB ram, this was causing even a simple
render to fail when the Linux OOM killer kicks in (often taking Blender
with it).  Passing the defaults as-is is much more reasonable; everything
is in KB.  This is noted in the online docs, the args file, and tooltip.

Discussed with ihsieh.