Add backend changes for viewport resolution multiplier
We now scale the resolution of the image to render for the viewport by the
new viewport resolution multiplier (which ihsieh just added).  Then we pass
the actual resolution of the viewport to the display driver when drawing
into the viewport and allow OpenGL to resample the render for us.  Note
that this commit requires a correspondingly updated build of the d_blender
display driver in prman in order to work.