Fix for RMAN-16122.
In the camera translator, split out the viewport resolution/screen window code to its own function,
so we can set them independently of the FOV/projection.

As to what was causing the issue, even though we've set screen window on
the camera, we seem to always be getting the default screenwindow. The workaround is to
set the screen window at the scene option level
when we have any projection changes.