Checkpoint. This commit should make it possible for mesh lights to have light filters, be light linked, and edited via the light mixer editor.
This also removes the light group editor.
The problem is trying to have a
different lightGroup for each instance
of a mesh light with the same material
network. We'll have to come up with
a different UI. For now, you'll have to
manually type in the lightGroup name
in the shader params.