Remove the open/close arrow that would open/close a page for our config properties.
The open/close
status was being stored as BoolProperty. Every time
the page was opened or close the BoolProperty
has to be updated and this would trigger as an object update in IPR.
Since Blender doesn't tell us what actually changed we
have to re-emit the whole geometry. Removing this
BoolProperty will save us from these false-positives.