Fix issues found when rendering the stirling model.
1. Make sure we don't consider
empties as new objects
2. Only remove/add instances when
there are any

We are still slower than Cycles
when we move the parent node
of the stirling. When we move the
parent node, Blender tells us that
each child node underneath the
parent has moved, rather than telling
us just the parent node move. We
should re-visit this to see if there's
a better way to handle this.