More WIP for stylized looks editor.
1. We now have a small beginnings
of stylized looks presets, which can
ben applied to the display filters
and the patterns.
2. Add some fixes to texture paths
and texture manager to allow for
relative paths.
3. Re-work the AOVs manager a
little bit. We now store all of the
scene's display channels on the
RendermanRenderLayerSettings
property group and then each AOV
references that list. This should
hopefully prevent name collisions.