A couple of UI updates.
* For our floating editors, workaround the issue where
clicking away from the window will make the editor disappear.
Clicking away from the window triggers a cancel, so if we detect
a cancel was done via a left mouse click, just call the opeartor again.
* Draw the txmanager in a panel again. One problem with using an operator,
we can't explicilty tell it to re-draw.
We keep the operator version around for opening the txmanager from a shading node.